Rename 'api_docs' package to just 'docs'
