+++ pyinaturalist/docs/emoji.py
