+++ pyinaturalist/docs/model_docs.py
