+++ pyinaturalist/docs/templates.py
